refactor(server): renamed main.go -> server.go
